- "success" and "error" dialogs; - refactorings.
